backend/requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dm-tree==0.1.8
exceptiongroup==1.2.0
fastapi==0.108.0
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.26.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.7
numpy==1.26.3
oauthlib==3.2.2
openai==1.6.1
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dotenv
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
# tensorflow==2.15.0.post1
# tensorflow-estimator==2.15.0
# tensorflow-io-gcs-filesystem==0.35.0
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor==2.4.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
Werkzeug==3.0.1
wrapt==1.14.1
